feat: Move authentication options to a separate file for better organization and maintainability
